Support inbound message moderations
Port in most of the message moderation rendering support from Cheogram
Android.

Things Cheogram Android does which this currently lacks require storing
in the DB for a message that it was moderated, such as:

* Showing nickname as "moderated"
* Hiding avatar on moderated messages
* Merging consecutive moderated messages into one

This also does not add support for a moderator to actually moderate
messages, only to show moderated messages as such if moderated by
another client.